requirements.txt
notion-client==2.2.1
python-dotenv==1.0.0
google-api-python-client==2.118.0